+++ src/main/java/com/sudurukbackback/modulecture/domain/auth/dto/request/UserLoginRequestDto.java
@@ -1,0 +1,4 @@
+package com.sudurukbackback.modulecture.domain.auth.dto.request;
+
+public class UserLoginRequestDto extends BaseAuthRequest {
+}
