+++ src/main/java/com/sudurukbackback/modulecture/domain/auth/dto/request/UserRegistrationRequestDto.java
@@ -1,4 +1,4 @@
-package com.sudurukbackback.modulecture.domain.user.dto.request;
+package com.sudurukbackback.modulecture.domain.auth.dto.request;
 
 import jakarta.validation.constraints.Size;
 import lombok.Data;
