--- src/main/java/com/sudurukbackback/modulecture/domain/user/dto/request/UserLoginRequestDto.java
@@ -1,4 +1,0 @@
-package com.sudurukbackback.modulecture.domain.user.dto.request;
-
-public class UserLoginRequestDto extends BaseAuthRequest {
-}
